refactor(direct-control): extract target candidates source
Move the target-candidates embedded App UI source into src/play/tactical/target-candidates.ts while keeping the public wrapper and command builder in the facade. Keep tactical helper logic local to the target-candidates owner rather than creating a broad shared tactical catalog.

Record the tactical read atom status in the support modularization OpenSpec workstream. This is source ownership proof only and does not claim live-game runtime proof.

Validation: bun run --cwd packages/civ7-direct-control test -- tactical-reads.test.ts; bun run --cwd packages/civ7-direct-control test; bun run --cwd packages/civ7-direct-control check; bun run --cwd packages/civ7-direct-control build; bun run openspec -- validate civ7-support-direct-control-modularization --strict; bun run resources:status; git diff --check.
