+++ openspec/changes/civ7-support-direct-control-modularization/tasks.md
@@ -129,6 +129,8 @@ runtime/direct-control claims.
         public wrapper in the facade.
   - [x] 4.7.3 Extract progress dashboard embedded source owner while keeping
         the public wrapper in the facade.
+  - [x] 4.7.4 Extract target-candidates embedded source owner while keeping the
+        public wrapper in the facade.
 - [ ] 4.8 Export stable types/constants only after module owners are defined.
   - [x] 4.8.1 Extract ComponentID primitive and direct-control error owner
         modules behind the existing package facade.
